productor/requirements.txt
paho-mqtt==1.6.1
numpy==2.0.1